# MetricNamer: avoid adding duplicate suffixes when UTF8 is allowed