@dashpole 2025-07-23
Btw, this performance penalty wold only occur if someone is using UTF8Allowed=true and withSuffixes=true, which should be rare (outside of the current Go SDK).