@dashpole 2025-07-23
Results are now ~2% better than before thanks to https://github.com/prometheus/otlptranslator/pull/48#discussion_r2226378374