@ArthurSens 2025-07-24
Any other concerns @ywwg? We have another release of the collector next Monday, I know it's kinda rushed but it would be nice to have a fix by then