@ArthurSens 2025-07-25
Spoke a bit offline, I'm taking the liberty to merge since we've got a release to do that requires this fix